pyproject.toml
[tool.poetry]
name = "problems"
version = "0.1.0"
description = "Python implementation of RFC 7870 \"Problem Details for HTTP APIs\""
authors = ["Berislav Lopac <berislav@lopac.net>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.7"
rfc3986 = {git = "https://git@github.com/python-hyper/rfc3986.git"}
jsonschema = "^3.2.0"
tinydb = "^3.15.2"
python-box = "^4.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.3.4"
black = "^19.10b0"
pytest-cov = "^2.8.1"
pytest-sugar = "^0.9.2"
flake8 = "^3.8.4"
tox = "^3.21.4"
mypy = "^0.800"
pytest-mypy = "^0.8.0"

[tool.coverage.run]
source = [ "src/problems/", ]
omit = [ "*/tests/*", "src/tests/*", ]

[tool.coverage.report]
skip_covered = true
show_missing = true
fail_under = 90

[tool.black]
line-length = 96
target-version = ['py37', 'py38']
verbose = false
skip-string-normalization = false

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
